+++ src/Batch/Batch/ComputeNodes/GetBatchComputeNodeRemoteLoginSettingCommand.cs
@@ -14,15 +14,12 @@
 
 using Microsoft.Azure.Batch;
 using Microsoft.Azure.Commands.Batch.Models;
-using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 using System.Management.Automation;
 using Constants = Microsoft.Azure.Commands.Batch.Utils.Constants;
 
 namespace Microsoft.Azure.Commands.Batch
 {
-    [GenericBreakingChange("Get-AzBatchRemoteLoginSettings alias will be removed in an upcoming breaking change release")]
     [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzurePrefix + "BatchRemoteLoginSetting", DefaultParameterSetName = Constants.IdParameterSet),OutputType(typeof(PSRemoteLoginSettings))]
-    [Alias("Get-AzBatchRemoteLoginSettings")]
     public class GetBatchComputeNodeRemoteLoginSettingCommand : BatchObjectModelCmdletBase
     {
         [Parameter(Position = 0, ParameterSetName = Constants.IdParameterSet, Mandatory = true,