+++ src/Batch/Batch/Pools/GetBatchPoolUsageMetricCommand.cs
@@ -16,14 +16,10 @@ using System;
 using Microsoft.Azure.Batch;
 using Microsoft.Azure.Commands.Batch.Models;
 using System.Management.Automation;
-using Constants = Microsoft.Azure.Commands.Batch.Utils.Constants;
-using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 
 namespace Microsoft.Azure.Commands.Batch
 {
-    [GenericBreakingChange("Get-AzBatchPoolUsageMetrics alias will be removed in an upcoming breaking change release")]
     [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzurePrefix + "BatchPoolUsageMetric"), OutputType(typeof(PSPoolUsageMetrics))]
-    [Alias("Get-AzBatchPoolUsageMetrics")]
     public class GetBatchPoolUsageMetrics : BatchObjectModelCmdletBase
     {
         [Parameter]